python-app/pyproject.toml
[project]
name = "python-app"
version = "0.1.0"
description = "RAG application using Gemini API, MariaDB vector database, and FAISS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "mariadb",
    "faiss-cpu",
    "sentence_transformers >=2.3.0",
    "numpy !=2.4.0",
    "python-dotenv",
]
